% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billonnage_non_regionalise_2015.R
\name{ABCD_DHP2015}
\alias{ABCD_DHP2015}
\title{Fonction interne qui prévoit la répartition par produits des arbres de 6 essences d'arbres feuillus à l'aide des
équations de Petro de
Havreljuk et al. (2015).}
\usage{
ABCD_DHP2015(data, type)
}
\arguments{
\item{data}{Un dataframe qui contient en ligne les arbres dont on veut prévoir
les rendements en produit à l'aide du module de billonnage Petro 2015.
Le dataframe doit contenir une colonne TigeID qui numérote individuellement chacune des lignes.
Doit contenir aussi les colonnes DHPcm et Espece, et optionnelement MSCR ou ABCD ou vigu0 et prod0.
Les équations ne s'appliquent qu'aux espèces: "ERS", "BOJ", "ERR", "BOP", "HEG", "CHR", les autres seront supprimés
Les équations ne s'appliquent qu'aux arbres avec un dhp>23, les autres seront supprimés}

\item{type}{"ABCD2015" pour utiliser les équations basées sur ABCD
"1234" pour utiliser les équations basées sur 1234
"MSCR" pour utiliser les équations basées sur MSCR
"DHP2015" pour utiliser les équations basées seulement sur le DHP}
}
\value{
Retourne un dataframe avec l'estimation du volume par classe de produit
         pour chacun des arbres "ERS", "BOJ", "ERR", "BOP", "HEG", "CHR" de 23 cm,
         colonnes: TigeID, type, F1, F2, F3, F4, P, DER
}
\description{
Fonction interne qui prévoit la répartition par produits des arbres de 6 essences d'arbres feuillus à l'aide des
équations de Petro de
Havreljuk et al. (2015).
}
\keyword{internal}
