% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billonnage_regionalise.R
\name{ABCD_DHP_region}
\alias{ABCD_DHP_region}
\title{Fonction interne qui prévoit la répartition par produits des arbres ERS et BOJ à l'aide des nouvelles
équations de Petro régionalisées de Havreljuk et al. (2025).}
\usage{
ABCD_DHP_region(data, type)
}
\arguments{
\item{data}{Un dataframe qui contient en ligne les arbres dont on veut prévoir
 les rendements en produit à l'aide du module de billonnage Petro
 régionalisé.
 Le dataframe doit contenir une colonne TigeID qui numérote individuellement chacune des lignes
 et une colonne, Espece, colonne eco qui contient le groupe de régions écologiques,
 une colonne DHPcm et optionnelement une colonne ABCD.
 Les équations ne s'appliquent qu'aux espèces: "ERS", "BOJ", les autres seront supprimés
 Les équations ne s'appliquent qu'aux arbres avec un dhp>23, les autres seront supprimés
 Les équations ne doivent êtres appliquées que sur les régions écologiques:
"1a", "2a", "2b", "2c", "3a", "3b", "3c", "3d", "4a", "4b", "4c", "4d", "4e", "4f", "4g","4h", "DU", "SV"}

\item{type}{"ABCD" pour utiliser les équations basées sur ABCD
"DHP" pour utiliser les équations basées seulement sur le DHP}
}
\value{
Retourne un dataframe avec l'estimation du volume par classe de produit
         pour chacun des arbres BOJ et ERS de 23 cm
         colonnes: TigeID, type, F1, F2, F3, F4, P, DER
}
\description{
Fonction interne qui prévoit la répartition par produits des arbres ERS et BOJ à l'aide des nouvelles
équations de Petro régionalisées de Havreljuk et al. (2025).
}
\keyword{internal}
