% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BillonnagePetro-package}
\alias{BillonnagePetro}
\alias{BillonnagePetro-package}
\title{BillonnagePetro: Équations de billonnage PETRO}
\description{
Équations pour estimer le volume des produits PETRO de 6 essences de feuillus avec les équations Havreljuk et al. (2015) ou avec les équations régionalisées pour ERS et BOJ de Havreljuk et al. (2025).
}
\keyword{internal}
