% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liste_arbres_ex}
\alias{liste_arbres_ex}
\title{Dataframe avec 6 arbres}
\format{
## `liste_arbres_ex`
A data frame with 6 lignes et 9 colonnes:
\describe{
  \item{Placette}{Identifiant de la placette}
  \item{reg_eco}{Code de la région écologique}
  \item{Espece}{Code d'essence de l'arbre}
  \item{DHPcm}{DHP de l'arbre(cm)}
  \item{vigu0}{Code de vigueur de l'arbre}
  \item{prod0}{Code de produit de l'arbre}
  \item{MSCR}{Code MSCR de l'arbre}
  \item{ABCD}{Code ABCD de l'arbre}
  \item{TigeID}{Identifiant unique de l'arbre}
}
}
\usage{
liste_arbres_ex
}
\description{
Dataframe avec 6 arbres et les informations nécessaires pour estimer le volume par produit Petro
}
\examples{
liste_arbres_ex
}
\keyword{datasets}
