% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBillonnagePetro.R
\name{SimBillonnagePetro}
\alias{SimBillonnagePetro}
\title{Fonction principale qui prévoit la répartition par produits des arbres de 6 essences d'arbres feuillus avec les équations Petro de Havreljuk et al. (2015)
ou avec les équations Petro régionalisées de Havreljuk et al. (2025).}
\usage{
SimBillonnagePetro(Data, type)
}
\arguments{
\item{Data}{Un dataframe qui contient en ligne les arbres dont on veut prévoir
les rendements en produit Petro.
Le dataframe doit contenir une colonne TigeID qui numérote individuellement chacune des lignes.
Doit aussi contenir les colonnes Espece et DHPcm
Pour type=DHP et type=ABCD, il faut la colonne eco ou reg_eco
Pour type=ABCD ou ABCD2015, il faut aussi une colonne ABCD
Pour type=MSCR, il faut aussi une colonne MSCR
Pour type=1234, il faut aussi les colonnes prod0 et vigu0
Les équations ne s'appliquent qu'aux arbres avec un dhp>23, les autres seront supprimés
Les équations ne s'appliquent qu'aux espèces: "ERS", "BOJ", "ERR", "BOP", "HEG", "CHR", les autres seront supprimés}

\item{type}{"DHP" pour utiliser les équations régionalisées de 2025 basées seulement sur le DHP
"ABCD" pour utiliser les équations régionalisées de 2025 basées sur ABCD
"1234" pour utiliser les équations de 2015 basées sur 1234
"MSCR" pour utiliser les équations de 2015 basées sur MSCR
"DHP2015" pour utiliser les équations de 2015 basées seulement sur le DHP
"ABCD2015" pour utiliser les équations de 2015 basées sur ABCD}
}
\value{
Retourne un dataframe avec l'estimation du volume par classe de produits
         pour chacun des arbres ERS, BOJ, ERR, BOP, HEG, CHR pour Petro 2015 ou chacun des arbres ERS et BOJ pour Petro 2025
}
\description{
Fonction principale qui prévoit la répartition par produits des arbres de 6 essences d'arbres feuillus avec les équations Petro de Havreljuk et al. (2015)
ou avec les équations Petro régionalisées de Havreljuk et al. (2025).
}
\details{
Les équations régionaliées (type DHP et ABCD) ne doivent êtres appliquées que sur les régions écologiques:
         "1a", "2a", "2b", "2c", "3a", "3b", "3c", "3d", "4a", "4b", "4c", "4d", "4e", "4f", "4g","4h", "DU", "SV"
}
\examples{
vol_billon <- SimBillonnagePetro(Data=liste_arbres_ex, type="ABCD")
}
